+++ src/bda/plone/productshop/behaviors.py
@@ -78,11 +78,6 @@ class IProductManualBehavior(model.Schema):
 class IProductBehavior(model.Schema):
     """Product behavior.
     """
-    item_number = schema.TextLine(
-        title=_(u'item_number_title', default=u'Item number'),
-        description=_(u'item_number_description',
-                      default=u'Item number of the product'),
-        required=False)
 
     image = NamedBlobImage(
         title=_(u'image_title', default=u'Product Image'),